+++ src/Controller/Admin/LearningMaterialsGroupController.php
@@ -11,6 +11,7 @@ namespace App\Controller\Admin;
 
 use App\Entity\Admin\LearningMaterialsGroup;
 use App\Form\Admin\LearningMaterialsGroupType;
+use App\Repository\Admin\LearningMaterialRepository;
 use App\Repository\Admin\LearningMaterialsGroupRepository;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
@@ -59,27 +60,71 @@ class LearningMaterialsGroupController extends AbstractController
         $learningMaterialsGroupInformation= new LearningMaterialsGroupRepository();
         $id = $learningMaterialsGroupInformation -> getQuantity();
         if($id>0) {
+            $info=true;
             for ($i = 0; $i < $id; $i++) {
                 $learningMaterialsGroup = $learningMaterialsGroupInformation->getLearningMaterialsGroup($i);
 
                 $tplArray[$i] = array(
                     'id' => $i,
                     'name_of_group' => $learningMaterialsGroup['name_of_group'],
-                    'exam_id' => $learningMaterialsGroup['exam_id'],
                 );
             }
         } else {
+            $info = false;
             $tplArray = array(
                 'id' => "",
                 'name_of_group' => "",
-                'exam_id' => "",
-
             );
         }
+        if( isset( $_SESSION['information'] ) && count( $_SESSION['information'] ) > 0  ) {
+            $infoDelete = $_SESSION['information'];
+        } else {
+            $infoDelete = "";
+        }
+        $_SESSION['information'] = array();
         return $this->render( 'LearningMaterialsGroupList.html.twig', array (
-            'data' => $tplArray
+            'data' => $tplArray,
+            'information' => $info,
+            'infoDelete' => $infoDelete
         ) );
     }
+    /**
+     * @param Request $request
+     * @param LearningMaterialsGroup $learningMaterialsGroup
+
+     * @return \Symfony\Component\HttpFoundation\RedirectResponse|\Symfony\Component\HttpFoundation\Response
+     * @Route("editLearningMaterialsGroup/{id}", name="editLearningMaterialsGroup")
+     */
+    public function editExam(Request $request, LearningMaterialsGroup $learningMaterialsGroup)
+    {
+        $learningMaterialsGroupInformation = new LearningMaterialsGroupRepository();
+        $groupId = (int)$request->attributes->get('id');
+        $infos = $learningMaterialsGroupInformation->getLearningMaterialsGroup($groupId);
+
+        $examInfoArray = array(
+            'name_of_group' => $infos['name_of_group'],
+        );
+
+        $form = $this->createForm(LearningMaterialsGroupType::class, $learningMaterialsGroup);
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            $exams = $form->getData();
+            $entityManager = $this->getDoctrine()->getManager();
+
+
+            $values = $learningMaterialsGroup->getAllInformation();
+
+            $learningMaterialsGroupInformation->update($values,$groupId);
+            return $this->redirectToRoute('learningMaterialsGroupList');
+        }
+        return $this->render('learningMaterialsGroupAdd.html.twig', [
+            'form' => $form->createView(),
+            'examInformation' =>$examInfoArray,
+        ]);
+    }
+
+
     /**
      * @param Request $request
      * @Route("/deleteGroup/{learningMaterialsGroup}", name="deleteGroup")
@@ -89,9 +134,16 @@ class LearningMaterialsGroupController extends AbstractController
     {
         $id = $request->attributes->get('learningMaterialsGroup');
         $repo = new LearningMaterialsGroupRepository();
-        $repo->delete($id);
-        //todo: nie usuwac gdy sa powiazania
-        //todo: wyswietlanie, gdy brak grup
+        $learningMaterialRepo = new LearningMaterialRepository();
+        $isMaterial = $learningMaterialRepo->getQuantity($id);
+        if($isMaterial){
+            $_SESSION['information'][] = array( 'type' => 'error', 'message' => 'The record cannot be deleted, there are links in the database');
+
+        } else {
+            $repo->delete($id);
+            $_SESSION['information'][] = array( 'type' => 'ok', 'message' => 'Successfully deleted');
+
+        }
         return $this->redirectToRoute('learningMaterialsGroupList');
     }
 }
