+++ src/Controller/Admin/QuestionController.php
@@ -11,6 +11,7 @@ namespace App\Controller\Admin;
 
 use App\Entity\Admin\Question;
 use App\Form\Admin\QuestionType;
+use App\Repository\Admin\AnswerRepository;
 use App\Repository\Admin\QuestionRepository;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
@@ -25,18 +26,14 @@ class QuestionController extends AbstractController
      */
     public function new(Request $request)
     {
-        $repository = $this->getDoctrine()->getRepository(Question::class);
         $question = new Question([]);
+        $repositoryQuestion = new QuestionRepository();
+
         $examId = $request->attributes->get('examId');
         $form = $this->createForm(QuestionType::class, $question);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            $data[0] = $request->request->get('content');
-            $data[1] = $request->request->get('max_answers');
-            $data[2] = $request->request->get('is_multichoice');
-            $data[3] = $request->request->get('is_file');
-
             $idExamValue = $request->attributes->get('examId');
 
             $question = $form->getData();
@@ -44,9 +41,15 @@ class QuestionController extends AbstractController
             $entityManager = $this->getDoctrine()->getManager();
 
             $values = $question->getAllInformation();
+            if($form['file']->getData()==NULL) {
+                $filename = "";
+            } else {
+                $file = $form['file']->getData();
+                $filename = $repositoryQuestion->uploadFile($file);
+            }
 
-            $repositoryExam = new QuestionRepository();
-            $repositoryExam->insert($idExamValue, $values);
+
+            $repositoryQuestion->insert($idExamValue, $values,$filename);
 
             return $this->redirectToRoute('questionList', [
                 'id' => $examId,
@@ -66,47 +69,56 @@ class QuestionController extends AbstractController
      * @return \Symfony\Component\HttpFoundation\RedirectResponse|\Symfony\Component\HttpFoundation\Response
      */
     public function questionListCreate(Request $request) {
+        $_SESSION['exam_id'] ="";
         $questionInformation= new QuestionRepository();
         $examId = $request->attributes->get('id');
 
         $idQuestion = $questionInformation -> getQuantity($examId);
 
         if($idQuestion>0) {
+            $info = true;
             for ($i = 0; $i < $idQuestion; $i++) {
                 $questions = $questionInformation->getQuestion($examId,$i);
+
                 if ($questions['is_multichoice'] == 1) {
                     $is_required = "true";
                 } else {
                     $is_required = "false";
                 }
-                if ($questions['is_file'] == 1) {
-                    $is_required_file = "true";
-                } else {
-                    $is_required_file = "false";
-                }
 
                 $tplArray[$i] = array(
-                    'id' => $i,
+                    'id' => $questions['id'],
                     'exam_id' => $questions['exam_id'],
                     'content' => $questions['content'],
                     'max_answers' => $questions['max_answers'],
                     'is_multichoice' => $is_required,
-                    'is_file' => $is_required_file
+                    'name_of_file' => $questions['name_of_file'],
                 );
             }
         } else {
+            $info = false;
             $tplArray = array(
-                'id' => 0,
-                'exam_id' => 0,
-                'content' => 0,
-                'max_answers' => 0,
-                'is_multichoice' => 0,
-                'is_file' => 0
+                'id' => '',
+                'exam_id' => '',
+                'content' => '',
+                'max_answers' => '',
+                'is_multichoice' => '',
+                'name_of_file' => '',
             );
         }
+
+    if( isset( $_SESSION['information'] ) && count( $_SESSION['information'] ) > 0  ) {
+        $infoDelete = $_SESSION['information'];
+    } else {
+        $infoDelete = "";
+    }
+        $_SESSION['information'] = array();
+
         return $this->render( 'questionList.html.twig', array (
             'data' => $tplArray,
-            'examId' => $examId
+            'examId' => $examId,
+            'information' => $info,
+            'infoDelete' => $infoDelete
         ) );
     }
     /**
@@ -117,16 +129,100 @@ class QuestionController extends AbstractController
     public function deleteQuestion(Request $request)
     {
         $examId = $request->attributes->get('exam');
-
         $questionId = $request->attributes->get('question');
-        $repo = new QuestionRepository();
 
-        $repo->delete($examId, $questionId);
-        //todo: redirect to questionList nie usuwa zapytania ktore jest jako 1
-        //todo: zapytanie czy chce usunac  gdy sa powiazane answers
+        $repo = new QuestionRepository();
+        $answerRepo = new AnswerRepository();
+
+        $question = $repo->getQuestion($examId,$questionId);
+        $filename= $question['name_of_file'];
+
+        $isAnswer = $answerRepo->getQuantity($examId,$questionId);
+        if($isAnswer !=0 ){
+            $_SESSION['information'][] = array( 'type' => 'error', 'message' => 'The record cannot be deleted, there are links in the database');
+
+        } else {
+            $repo->delete($examId, $questionId);
+            $repo->deleteFile($filename);
+            $_SESSION['information'][] = array( 'type' => 'ok', 'message' => 'Successfully deleted');
+        }
         return $this->redirectToRoute('questionList', [
             'id' => $examId,
         ]);
     }
-    //todo: edit question
+
+    /**
+     * @param Request $request
+     * @param Question $question
+
+     * @return \Symfony\Component\HttpFoundation\RedirectResponse|\Symfony\Component\HttpFoundation\Response
+     * @Route("editQuestion/{exam_id}/{id}", name="editQuestion")
+     */
+    public function editQuestion(Request $request, Question $question)
+    {
+        $questionInformation = new QuestionRepository();
+        $examId = $request->attributes->get('exam_id');
+        $questionId = $request->attributes->get('id');
+
+        $_SESSION['exam_id'] = $examId;
+
+        $questions = $questionInformation->getQuestion($examId, $questionId);
+        $filenameFromDatabase = $questions['name_of_file'];
+
+        $questionInfoArray = array(
+            'content' => $questions['content'],
+            'max_answers' => $questions['max_answers'],
+            'is_multichoice' => $questions['is_multichoice'],
+            'name_of_file' => $questions['name_of_file'],
+        );
+
+        $form = $this->createForm(QuestionType::class, $question);
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+         /*   $exams = $form->getData();
+            $entityManager = $this->getDoctrine()->getManager();
+            $examValue = $request->attributes->get('id');*/
+            $filename = $filenameFromDatabase;
+
+            $values = $question->getAllInformation();
+            if($form['file']->getData()== NULL) {
+                $filename = $filenameFromDatabase;
+            } else {
+                $file = $form['file']->getData();
+                $filename=$questionInformation->updateFile($file,$filename);
+            }
+
+            $questionInformation->update($values,$examId, $questionId,$filename);
+
+            return $this->redirectToRoute('questionList', [
+                'id' => $questionId,
+            ]);
+        }
+        return $this->render('questionAdd.html.twig', [
+            'form' => $form->createView(),
+            'examInformation' =>$questionInfoArray,
+            'examId' => $examId
+        ]);
+    }
+
+
+    /**
+     * @param Request $request
+     * @Route("downloadFile/{exam}/{question}", name="downloadFile")
+     * @return \Symfony\Component\HttpFoundation\Response
+     */
+    public function downloadFile(Request $request)
+    {
+        $examId = $request->attributes->get('exam');
+
+        $questionId = $request->attributes->get('question');
+        $repo = new QuestionRepository();
+        $question = $repo->getQuestion($examId,$questionId);
+        $filename = $question['name_of_file'];
+        $url = "gs://examservicedatabase.appspot.com/" . $filename;
+        $file_name = basename($url);
+
+
+    }
 }
